feat: Add createProp function to main.js
The commit adds a new function called `createProp` to the `main.js` file in the `usefull-scripts` directory. This function is responsible for creating a proposal using the `ProposalsFactory` and `getContracts` utilities. It retrieves the necessary contracts, creates a proposal with default parameters, and then calls the `create` action on the `proposals` contract. The result of the action is logged to the console.

This addition enhances the functionality of the script by allowing users to easily create proposals.

Refactor: Remove redundant code in proposals::reset()
Fix linting issues
Update test configuration
Update error log
Remove unused code
Bug fix related to dereference of end iterator and dates
Comment unused code - fix linting