feat: Add deploy script for EOS contracts
This commit adds a new script called `deploy.sh` to the `usefull-scripts` directory. The script is used to deploy EOS contracts to the network. It takes two arguments: the account name and the contract name. The script checks if a compiled folder exists for the contract, and if so, it sets the contract using the `cleos` command. This script will be useful for developers who want to easily deploy their contracts.

Refactor: Remove redundant code in proposals::reset()
Fix linting issues
Update test configuration
Update error log
Remove unused code
Bug fix related to dereference of end iterator and dates
Comment unused code - fix linting